# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Petshop-api Documentation",
        "description": "petshop api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.0.107:8000",
            "description": "local server"
        },
        {
            "url": "http://staging.example.com",
            "description": "staging server"
        },
        {
            "url": "http://example.com",
            "description": "production server"
        }
    ],
    "paths": {
        "/lids": {
            "get": {
                "tags": [
                    "Lids"
                ],
                "summary": "Get list of lids",
                "description": "Returns list of lids",
                "operationId": "getLidsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Lids not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Lids"
                ],
                "summary": "Store new lid",
                "description": "Store new lid",
                "operationId": "storeLid",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the lid",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phone": {
                                        "description": "Phone number of the lid",
                                        "type": "string",
                                        "example": "998901234567"
                                    },
                                    "address": {
                                        "description": "Address of the lid",
                                        "type": "string",
                                        "example": "Tashkent"
                                    },
                                    "comment": {
                                        "description": "Comment of the lid",
                                        "type": "string",
                                        "example": "Comment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Lid not found"
                    }
                }
            }
        },
        "/lids/{lid}": {
            "put": {
                "tags": [
                    "Lids"
                ],
                "summary": "Update existing lid",
                "description": "Update existing lid",
                "operationId": "updateLid",
                "parameters": [
                    {
                        "name": "lid",
                        "in": "path",
                        "description": "ID of the lid",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the lid",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phone": {
                                        "description": "Phone number of the lid",
                                        "type": "string",
                                        "example": "998901234567"
                                    },
                                    "address": {
                                        "description": "Address of the lid",
                                        "type": "string",
                                        "example": "Tashkent"
                                    },
                                    "comment": {
                                        "description": "Comment of the lid",
                                        "type": "string",
                                        "example": "Comment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Lid not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Lids",
            "description": "Lids"
        }
    ]
}